Revampt some JWT stuff.
Instead of stuffing the path into the token, it's now duplicated.
This is less efficient but makes it more obvious that something be wrong.

We're also now more strict about paths that don't end with a `/`.
It needs to be improved, but these are not treated as prefixes any longer.