Add an ANNOUNCE_INIT message.
This message is the first reply received from the remote,
acting kind of like an ANNOUNCE_OK. It lets the caller know all
of the broadcasts currently available.

This is used to ensure that we have received all remote broadcasts
before the connection is considered established. This is purely a library
thing, the protocol itself does not require a client to know all broadcasts.
I just think it's nice when a client knows everything available at all times.

This will require a new protocol version unfortunately.