Merge Unique into Signal.
I was copying other JS libraries by using === to detect changes and avoid spurious updates.
However, this is such a frequent cause of bugs that I added log messages for it.

I keep trying to avoid dequal for performance reasons, but maybe I'm just being silly.
In Rust we would check for changes using PartialEq which would work like dequal.

So this is going with the "slower" but safer approach until benchmarks prove otherwise.